+++ lib/pinput.dart
@@ -1,3 +1,5 @@
 library pinput;
 
 export 'src/pinput.dart';
+
+export 'package:smart_auth/smart_auth.dart';